+++ src/commands/mod.rs
@@ -19,4 +19,5 @@ pub mod send;
 pub mod site;
 pub mod stack;
 pub mod task;
+pub mod task_publish;
 pub mod task_run;